+++ modules/node.sh
@@ -2,20 +2,24 @@
 # node.sh
 # Install nvm, the latest LTS version of Node and the latest actual version of Node..
 
-echo ""
+echo
 echo "---------------------------------------------------------------"
 echo "| Installing Latest Node and the LTS (default).               |"
 echo "---------------------------------------------------------------"
-echo ""
+echo
 
 echo "## Setting up NVM (Node Version Manager) ##"
 echo >> "$shell_rc"
 echo "# Set up NVM" >> "$shell_rc"
-curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.4/install.sh | bash
+if ! run_downloaded_installer nvm \
+  https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.6/install.sh \
+  nvm-installer.sh bash; then
+  return 1
+fi
 
 NVM_DIR="$([ -z "${XDG_CONFIG_HOME-}" ] && printf %s "${HOME}/.nvm" || printf %s "${XDG_CONFIG_HOME}/nvm")"
 export NVM_DIR
-[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh" # This loads nvm
+[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"                   # This loads nvm
 [[ -r $NVM_DIR/bash_completion ]] && \. "$NVM_DIR/bash_completion" # This loads nvm bash_completion
 
 # set up some default packages to always install - these 2 allow the use of a