+++ modules/python.sh
@@ -3,72 +3,71 @@
 # 'uv' and 'Poetry' for dependency management and 'pipx' for managing global
 # python packages. Python 2 is no longer installed by default, but can be
 # installed using 'pyenv' if needed.
-echo ""
+echo
 echo "---------------------------------------------------------------"
 echo "| Installing Python 3, uv, Poetry and PipX.                   |"
 echo "---------------------------------------------------------------"
-echo ""
+echo
 
 # we still install pyenv, though I now tend to use `uv` for python version
 # management too. Pyenv is still handy to have an updated global version
 # available.
-curl -L https://raw.githubusercontent.com/pyenv/pyenv-installer/master/bin/pyenv-installer | bash
-# install a couple of plugins...
-git clone https://github.com/yyuu/pyenv-pip-migrate.git ~/.pyenv/plugins/pyenv-pip-migrate
-git clone https://github.com/yyuu/pyenv-ccache.git ~/.pyenv/plugins/pyenv-ccache
-git clone https://github.com/jawshooah/pyenv-default-packages.git ~/.pyenv/plugins/pyenv-default-packages
+if run_downloaded_installer pyenv \
+  https://raw.githubusercontent.com/pyenv/pyenv-installer/master/bin/pyenv-installer \
+  pyenv-installer.sh bash; then
+  # install a couple of plugins...
+  git clone https://github.com/yyuu/pyenv-pip-migrate.git ~/.pyenv/plugins/pyenv-pip-migrate
+  git clone https://github.com/yyuu/pyenv-ccache.git ~/.pyenv/plugins/pyenv-ccache
+  git clone https://github.com/jawshooah/pyenv-default-packages.git ~/.pyenv/plugins/pyenv-default-packages
 
-# Set up a default-packages file for python libraries to install with each new
-# python or venv ... for now, just a few that allow my vscode settings to work
-# easier. Will probably deprecate this later, as I now use 'Poetry' for
-# dependency management.
-cat <<- EOF > ~/.pyenv/default-packages
+  # Set up a default-packages file for python libraries to install with each new
+  # python or venv ... for now, just a few that allow my vscode settings to work
+  # easier. Will probably deprecate this later, as I now use 'Poetry' for
+  # dependency management.
+  cat <<- EOF > ~/.pyenv/default-packages
 wheel
 ruff
 ipython
 EOF
 
-if ! grep -qc 'pyenv init' "$shell_rc" ; then
-  echo "## Adding pyenv to $shell_rc ##"
-  echo >> "$shell_rc"
-  echo "# Set up Pyenv" >> "$shell_rc"
-  echo 'export PYENV_ROOT="$HOME/.pyenv"' >> "$shell_rc"
-  echo 'command -v pyenv >/dev/null || export PATH="$PYENV_ROOT/bin:$PATH"' >> "$shell_rc"
-  echo "eval \"\$(pyenv init - $shell_type)\"" >> "$shell_rc"
-  echo 'eval "$(pyenv virtualenv-init -)"' >> "$shell_rc"
-fi
-# run the above locally to use in this shell
-export PYENV_ROOT="$HOME/.pyenv"
-command -v pyenv >/dev/null || export PATH="$PYENV_ROOT/bin:$PATH"
-eval "$(pyenv init -)"
-eval "$(pyenv virtualenv-init -)"
+  if ! grep -q 'pyenv init' "$shell_rc"; then
+    echo "## Adding pyenv to $shell_rc ##"
+    {
+      echo
+      echo "# Set up Pyenv"
+      echo 'export PYENV_ROOT="$HOME/.pyenv"'
+      echo 'command -v pyenv > /dev/null || export PATH="$PYENV_ROOT/bin:$PATH"'
+      echo "eval \"\$(pyenv init - $shell_type)\""
+      echo 'eval "$(pyenv virtualenv-init -)"'
+    } >> "$shell_rc"
+  fi
+  # run the above locally to use in this shell
+  export PYENV_ROOT="$HOME/.pyenv"
+  command -v pyenv > /dev/null || export PATH="$PYENV_ROOT/bin:$PATH"
+  eval "$(pyenv init -)"
+  eval "$(pyenv virtualenv-init -)"
 
-pyenv install 3.14
-pyenv global 3.14
-# now update 'pip' to the latest version ...
-python3 -m pip install --upgrade pip
-
-# add the .local/bin to the path if it isn't already there...
-if ! grep -qc '/.local/bin' "$shell_rc" ; then
-  echo 'export PATH="$HOME/.local/bin:$PATH"' >> "$shell_rc"
+  pyenv install 3.14
+  pyenv global 3.14
+  # now update 'pip' to the latest version ...
+  python3 -m pip install --upgrade pip
 fi
-# run the above locally to use in this shell
-export PATH="$HOME/.local/bin:$PATH"
 
 # Install 'poetry' for packaging and dependency management.
-curl -sSL https://install.python-poetry.org | python3 -
-
-# Tell poetry to create the venvs in the local project folder not global
-# This is optional, but I find it helps to cut down on the amount of random
-# venvs on your system, and keeps things obviously grouped.
-poetry config virtualenvs.in-project true
+if run_downloaded_installer poetry https://install.python-poetry.org \
+  poetry-installer.py python3; then
+  # Tell poetry to create the venvs in the local project folder not global
+  # This is optional, but I find it helps to cut down on the amount of random
+  # venvs on your system, and keeps things obviously grouped.
+  poetry config virtualenvs.in-project true
+fi
 
 # install pipx for managing global python packages
 python3 -m pip install --user pipx
 pipx ensurepath
 
 # add autocompletion for pipx
-if ! grep -qc 'pipx' "$shell_rc" ; then
+if ! grep -q 'pipx' "$shell_rc"; then
   echo 'eval "$(register-python-argcomplete pipx)"' >> "$shell_rc"
 fi
 
@@ -77,12 +76,13 @@ eval "$(register-python-argcomplete pipx)"
 
 # install 'uv' for python project and version management (replaces poetry). This
 # can also replace 'pipx' but we'll keep both available for now.
-curl -LsSf https://astral.sh/uv/install.sh | sh
-
-# prepare the uv config file if it does not exist
-mkdir -p "$HOME/.config/uv"
-touch "$HOME/.config/uv/uv.toml"
-if ! grep -qc 'python-preference' "$HOME/.config/uv/uv.toml"; then
-  # only use uv's own python versions, not any pyenv
-  echo 'python-preference="only-managed"' >> "$HOME/.config/uv/uv.toml"
+if run_downloaded_installer uv https://astral.sh/uv/install.sh \
+  uv-installer.sh sh; then
+  # prepare the uv config file if it does not exist
+  mkdir -p "$HOME/.config/uv"
+  touch "$HOME/.config/uv/uv.toml"
+  if ! grep -q 'python-preference' "$HOME/.config/uv/uv.toml"; then
+    # only use uv's own python versions, not any pyenv
+    echo 'python-preference="only-managed"' >> "$HOME/.config/uv/uv.toml"
+  fi
 fi