+++ modules/rust.sh
@@ -1,12 +1,15 @@
 #!/usr/bin/env bash
 # rust.sh Install the latest version of 'Rust' using 'Rustup'.
-echo ""
+echo
 echo "---------------------------------------------------------------"
 echo "| Installing Rust and related tools.                          |"
 echo "---------------------------------------------------------------"
-echo ""
+echo
 
-curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y
+if ! run_downloaded_installer rustup https://sh.rustup.rs \
+  rustup-installer.sh sh -y; then
+  return 1
+fi
 
 # add the .cargo/bin to the path for this shell in case we want to use it
 # later in the script.
@@ -16,7 +19,12 @@ export PATH="$HOME/.cargo/bin:$PATH"
 rustup component add clippy
 rustup component add rustfmt
 
-cargo install cargo-edit # upgrade dependencies from the CLI
-cargo install cargo-make # useful task runner
+# add 'cargo-binstall' to speed up later installs.
+if ! run_downloaded_installer cargo-binstall \
+  https://raw.githubusercontent.com/cargo-bins/cargo-binstall/main/install-from-binstall-release.sh \
+  cargo-binstall-installer.sh bash; then
+  :
+fi
 
-cargo install bob-nvim # neovim version manager
+install_with_cargo_binstall cargo-edit # upgrade dependencies from the CLI
+install_with_cargo_binstall cargo-make # useful task runner